% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{doc_md5}
\alias{doc_md5}
\title{Calculate the MD5 digest for a file}
\usage{
doc_md5(f)
}
\arguments{
\item{f}{Path to a file}
}
\description{
Calculate the MD5 digest for a file
}
\details{
A simple wrapper over \link[digest]{digest}, with \code{file = TRUE}
and \code{errormode = "warn"}.
}
\examples{
\dontrun{
  md5 <- doc_md5("~/Downloads/test.pdf")
}
}

