% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecosdata-package.R
\docType{package}
\name{ecosdata}
\alias{ecosdata}
\alias{ecosdata-package}
\title{ecosdata: Get data and documents from ECOS}
\description{
ecosdata: Get data and documents from ECOS
}
\section{About}{


The U.S. Fish and Wildlife Service provides documents (and some limited data
about the documents) on ECOS, \url{http://ecos.fws.gov}. But there is no API
for the documents, so we have to scrape the contents. Builds on
\link[ecosscraper]{ecosscraper} to do the "analysis" of scraping ECOS and
storing the information in a central location.
}

