% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{file_info}
\alias{file_info}
\title{Slightly improved \link[base]{file.info}}
\usage{
file_info(...)
}
\arguments{
\item{...}{Paths to one or more files to gather info}
}
\description{
Slightly improved \link[base]{file.info}
}
\details{
Move the file name from rowname and into \code{$path} of
\link[tidyr]{data_frame}.
}
\examples{
\dontrun{
  info <- file_info("~/Downloads/test.pdf")
}
}

