% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{GET_page}
\alias{GET_page}
\title{Get the HTML version of the page at a URL}
\usage{
GET_page(url, pause = TRUE)
}
\arguments{
\item{url}{The URL to fetch}

\item{pause}{Pause for 0-3s before fetching? [default = TRUE]}
}
\description{
Function description; defaults to title if left blank
}
\examples{
An example to be run during package build
\dontrun{
  page <- GET_page(TECP_table$Species_Page[1])
}
}

