% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{get_species_baseline}
\alias{get_species_baseline}
\title{Get baseline hashes of species' ECOS pages}
\usage{
get_species_baseline(urls, species, pause = FALSE)
}
\arguments{
\item{urls}{The species page URLs to fetch}

\item{species}{The scientific names of the species to process}

\item{pause}{Whether to pause 0.5-3s between page fetches [default = FALSE]}
}
\description{
Get baseline hashes of species' ECOS pages
}
\examples{
\dontrun{
   sp_base <- get_species_baseline(TECP_table$Species_Page[1:3],
                                   TECP_table$Scientific_Name[1:3],
                                   FALSE)
}
}

