% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TECP_baseline.R
\name{get_TECP_baseline}
\alias{get_TECP_baseline}
\title{Get a baseline hash of the ECOS species report table (TECP_table)}
\usage{
get_TECP_baseline()
}
\value{
A list with two elements, \describe{
  \item{TECP_table}{The table of T, E, candidate, and proposed species}
  \item{TECP_summary}{A data.frame with summary information from the scrape}
}
}
\description{
Fetches the table at \url{https://goo.gl/tZRu5o}, and returns both the table
as a data.frame and a summary of the scrape including the MD5 of the table
(same structure as returned by \link[ecosscraper]{get_species_page_summary}).
}

