% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize.R
\name{species_page_md5}
\alias{species_page_md5}
\title{Calculate the MD5 digest for an ECOS page}
\usage{
species_page_md5(f)
}
\arguments{
\item{f}{Path to a species' ECOS page (either local or URL)}
}
\description{
Calculate the MD5 digest for an ECOS page
}
\note{
Apparently, FWS does not serve the same version of a species' page up
twice in a row. Instead, the same information will be presented in different
orders. We have to use \code{strsplit} along with \link[stringr]{str_trim} to
get clean lines, then sort before doing the MD5 hash.
}
\examples{
\dontrun{
  md5 <- species_page_md5("https://goo.gl/kvIcH9")
}
}

